Check autorun script configure_zaprets.sh

+++ awg_config.sh
@@ -342,9 +342,13 @@ printf  "\033[32;1mStop and disabled service 'youtubeUnblock' and 'ruantiblock'.
 manage_package "youtubeUnblock" "disable" "stop"
 manage_package "ruantiblock" "disable" "stop"
 
-grep -v "0 4 \* \* \* wget -O - $URL/configure_zaprets.sh | sh" /etc/crontabs/root > /etc/crontabs/temp
-cp -f "/etc/crontabs/temp" "/etc/crontabs/root"
-rm -f "/etc/crontabs/temp"
+str=$(grep -i "0 4 \* \* \* wget -O - $URL/configure_zaprets.sh | sh" /etc/crontabs/root)
+if [ ! -z "$str" ]
+then
+	grep -v "0 4 \* \* \* wget -O - $URL/configure_zaprets.sh | sh" /etc/crontabs/root > /etc/crontabs/temp
+	cp -f "/etc/crontabs/temp" "/etc/crontabs/root"
+	rm -f "/etc/crontabs/temp"
+fi
 
 printf  "\033[32;1mRestart firewall and network...\033[0m\n"
 service firewall restart
